Fixed so if there is no logger an exception shouldn't be thrown (stops having to force a logger on the users)
Changed Health endpoint to not use as much linq and to not use WaitAll (blocking) and instead use WhenAll which is async

Moved the response to the bottom and added a summary object with basic stats that we can add to

Added a counter tracker with just HTTP Responses + UpTime. Made a stats per thread as it is small with a summary across the objects

Added logging to the routing, added tree compression.

Removed the FakeHealthRouter as it is no longer needed

 Removed the default logger (now up to the user)

Added the CurrentState object to the Dispose